util 추가 및 출력 포맷 JSON으로 변경
