--- prompts/one_shot_prompt.yaml
@@ -1,106 +1,0 @@
-gemba_mqm_prompt: |
-  Given the source segment and machine translation, identify error types in the translation and classify them.
-  The error types are:
-  - Accuracy:
-    * Addition: The translation includes information not present in the source.
-    * Mistranslation: The translation does not accurately represent the source.
-    * Omission: The translation is missing content from the source.
-    * Untranslated text: Text is left untranslated.
-  - Fluency:
-    * Character encoding: There are garbled characters or mojibake.
-    * Grammar: There are issues with grammar.
-    * Inconsistency: The translation is inconsistent in style, terminology, or otherwise.
-    * Punctuation: There are issues with punctuation.
-    * Register: The language is of an inappropriate register.
-    * Spelling: There are misspellings.
-  - Style: The translation style is not appropriate.
-  - Terminology: The translation has terminology problems.
-  - Other: Any other issues.
-  - Non-translation: A non-translation issue, e.g., the source was corrupted.
-  - No-error: No error found in this segment.
-
-  Classify each error as one of:
-  - Critical: Errors that render the translation completely wrong, misleading, or incomprehensible.
-  - Major: Errors that impact comprehension but do not completely mislead or confuse the user.
-  - Minor: Errors that do not impact comprehension.
-
-  Calculate the MQM score as: 100 - (25 x Critical + 5 x Major + 1 x Minor).
-
-  Here is one example:
-
-  Example:
-  English source: So we attach this by pressing Ctrl+J.
-  Korean translation: 그래서 Ctrl+J를 눌러서 이걸 붙여주는 거죠.
-  MQM annotations:
-  Critical: 
-  No-error
-  Major: 
-  Mistranslation: '붙여주는' (attaching) vs. '연결' (connecting)
-  Minor: 
-  Register: "거죠" is too informal compared to a more appropriate "합니다"
-  MQM score: 94
-
-  Now evaluate this translation:
-
-  English source: {source_text}
-  Korean translation: {translation}
-
-  Identify and classify errors, then calculate the final MQM score.
-
-ea_prompt: |
-  Given the source segment and machine translation, identify error types in the translation and classify them.
-  The error types are:
-  - Accuracy:
-    * Addition: The translation includes information not present in the source.
-    * Mistranslation: The translation does not accurately represent the source.
-    * Omission: The translation is missing content from the source.
-    * Untranslated text: Text is left untranslated.
-  - Fluency:
-    * Character encoding: There are garbled characters or mojibake.
-    * Grammar: There are issues with grammar.
-    * Inconsistency: The translation is inconsistent in style, terminology, or otherwise.
-    * Punctuation: There are issues with punctuation.
-    * Register: The language is of an inappropriate register.
-    * Spelling: There are misspellings.
-  - Style: The translation style is not appropriate.
-  - Terminology: The translation has terminology problems.
-  - Other: Any other issues.
-  - Non-translation: A non-translation issue, e.g., the source was corrupted.
-  - No-error: No error found in this segment.
-
-  Classify each error as one of:
-  - Critical: Errors that render the translation completely wrong, misleading, or incomprehensible.
-  - Major: Errors that impact comprehension but do not completely mislead or confuse the user.
-  - Minor: Errors that do not impact comprehension.
-  
-  Calculate the MQM score as: 100 - (10 x Major + 2 x Minor).
-
-  Here is one example:
-
-  Q:
-  English source: So we attach this by pressing Ctrl+J.
-  Reference: Ctrl+J를 눌러서 이것을 연결합니다.
-  Korean translation: 그래서 Ctrl+J를 눌러서 이걸 붙여주는 거죠.
-
-  Based on the given source and reference, identify the major and minor errors in this
-  translation. Note that Major errors refer to actual translation or grammatical errors, and
-  Minor errors refer to smaller imperfections, and purely subjective opinions about the
-  translation.
-
-  A:
-  Major errors:
-  (1) "붙여주는" - Mistranslation
-  Minor errors:
-  (1) "이걸" and "거죠" - Register
-  MQM score = 100 - (10 + 2) = 86
-
-  Q:
-  English source: {source_text}
-  Korean translation: {translation}
-
-  Based on the given source and reference, identify the major and minor errors in this
-  translation. Note that Major errors refer to actual translation or grammatical errors, and
-  Minor errors refer to smaller imperfections, and purely subjective opinions about the
-  translation.
-  
-  A:
