+++ results/gpt4.1/ea_prompt_gpt-4.1_fewshot_correlation.json
@@ -1,0 +1,7 @@
+{
+    "prompt_type": "ea_prompt",
+    "model": "gpt-4.1",
+    "kendall_tau": 0.2283302163985783,
+    "p_value": 7.8619979408082e-06,
+    "sample_size": 250
+}
