+++ results/gpt4o/ea_prompt_gpt-4o_fewshot_correlation.json
@@ -1,0 +1,7 @@
+{
+    "prompt_type": "ea_prompt",
+    "model": "gpt-4o",
+    "kendall_tau": 0.34990732624769405,
+    "p_value": 1.4234529596149219e-11,
+    "sample_size": 250
+}
