+++ utils/logger_util.py
@@ -1,0 +1,18 @@
+import logging
+
+# 로깅 설정
+def setup_logger():
+    """로깅 설정 초기화"""
+    logger = logging.getLogger()
+    logger.setLevel(logging.INFO)
+    
+    # 콘솔 핸들러 설정
+    console_handler = logging.StreamHandler()
+    console_handler.setLevel(logging.INFO)
+    console_formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    console_handler.setFormatter(console_formatter)
+    logger.addHandler(console_handler)
+
+    logging.getLogger("httpx").setLevel(logging.WARNING)
+    
+    return logger
